+++ opteryx/engine/planner/operations/limit_node.py
@@ -22,12 +22,14 @@ from typing import Iterable
 from pyarrow import Table, concat_tables
 
 from opteryx.engine.planner.operations.base_plan_node import BasePlanNode
-from opteryx.engine.query_statistics import QueryStatistics
+from opteryx.engine import QueryDirectives, QueryStatistics
 from opteryx.exceptions import SqlError
 
 
 class LimitNode(BasePlanNode):
-    def __init__(self, statistics: QueryStatistics, **config):
+    def __init__(
+        self, directives: QueryDirectives, statistics: QueryStatistics, **config
+    ):
         self._limit = config.get("limit")
 
     @property